interface refactor start
